[2차 세미나 과제] - UI View, ViewController 동작 구현
