+++ GO_SOPT_Seminar_Assingment/SecondSeminar/Assignment_TVING/View/AddNickNameBottomSheetUIView.swift
@@ -83,10 +83,6 @@ class AddNickNameBottomSheetUIView: UIView {
         )
     }
     
-    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        self.endEditing(true)
-    }
-    
 }
 
 private extension AddNickNameBottomSheetUIView {
