+++ GO_SOPT_Seminar_Assingment/SecondSeminar/Assignment_TVING/WelcomeViewController.swift
@@ -7,66 +7,26 @@
 
 import UIKit
 
-class WelcomeViewController: UIViewController {
+class WelcomeViewController: BaseViewController {
     
-    private let mainImgView = UIImageView().then {
-        $0.image = UIImage(named: "tiving_logo")
-    }
+    private let mainView = WelcomeView()
     
-    private let welcomeLabel = UILabel().then {
-        $0.text = "님\n반가워요!"
-        $0.numberOfLines = 2
-        $0.font = .tvingSemiBold(ofSize: 23)
-        $0.textColor = .white
-        $0.textAlignment = .center
-    }
-    
-    private lazy var goToMainBtn = UIButton().then {
-        $0.setTitle("메인으로", for: .normal)
-        $0.titleLabel?.font = .tvingMedium(ofSize: 14)
-        $0.titleLabel?.textAlignment = .center
-        $0.backgroundColor = .tvingRed
-        $0.layer.cornerRadius = 3
-        $0.addTarget(self, action: #selector(tappedGoToMainBtn), for: .touchUpInside)
+    private func target() {
+        mainView.goToMainBtn.addTarget(self, action: #selector(tappedGoToMainBtn), for: .touchUpInside)
     }
 
+    override func loadView() {
+        self.view = mainView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        style()
-        setLayout()
+        target()
     }
-    
 }
 
 extension WelcomeViewController {
-    
-    func style() {
-        view.backgroundColor = .black
-        self.navigationItem.setHidesBackButton(true, animated: true)
-    }
-    
-    func setLayout() {
-        
-        view.addSubviews(mainImgView, welcomeLabel, goToMainBtn)
-        
-        mainImgView.snp.makeConstraints {
-            $0.height.equalTo(mainImgView.snp.width).multipliedBy(0.56)
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(14)
-            $0.leading.trailing.equalToSuperview()
-        }
-        welcomeLabel.snp.makeConstraints {
-            $0.height.equalTo(75)
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(292)
-            $0.leading.equalToSuperview().inset(75)
-            $0.trailing.equalToSuperview().inset(74)
-        }
-        goToMainBtn.snp.makeConstraints {
-            $0.height.equalTo(52)
-            $0.leading.trailing.equalToSuperview().inset(20)
-            $0.bottom.equalToSuperview().inset(66)
-        }
-    }
     
     @objc
     func tappedGoToMainBtn() {
